Hotfix of annotation behavior.
If an annotation is displayed and the displayed data of the histogram is changed, now the annotation is correctly updated to show the fit results of the new data.

Minor bugfix:
Also, an error handling was implemented to the import function of the image data.
If the 'Forcemap' folder, created to import the force-map data of JPK maps, can not be deleted due to access restrictions the program no longer terminates with an error.
Instead, just a warning is displayed but the program continues.