Fix of the calculation of the "half angle to edge" in case the user enters the angle as "half angle to face".
The wrong calculation just multiplied the entered angle with sqrt(2).

But the factor of sqrt(2) is not for the angle valuse, it is just valid for the tan() of the angles.
If a = "half angle to edge" and b = "half angle to face", the correct relation is:
tan(a) = sqrt(2) *  tan(b)

Thus, the calculation of a is:
a = atan(sqrt(2)*tan(b))