Improved the import function of the "Sader Calibraiton GUI" to allow to import amplitude Sweeps with any unit and not just with "m".
Implemented the "new" Sader Method for certain Cantilevers, selectable via the new drop down menu.
This is for now deactivated, because the results of this "new" Sader Method is at the moment very different to other Methods.
Befor useres should use this, it needs still some clarification and remains deactivated.